+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/PageLayout/DefaultPageLayout.constants.js
@@ -24,10 +24,6 @@ export const DefaultPageLayout: PageLayoutConfig = {
     leftColumn: [
         {
             type: WidgetTypes.COMPONENT,
-            name: 'TwoEventWorkspace',
-        },
-        {
-            type: WidgetTypes.COMPONENT,
             name: 'EditEventWorkspace',
         },
     ],