+++ src/core_modules/capture-core/components/WidgetTwoEventWorkspace/WidgetHeader/WidgetHeader.container.js
@@ -1,0 +1,63 @@
+// @flow
+import React, { type ComponentType } from 'react';
+import { spacersNum } from '@dhis2/ui';
+import { withStyles } from '@material-ui/core/';
+import { EnrollmentPageKeys }
+    from '../../Pages/common/EnrollmentOverviewDomain/EnrollmentPageLayout/DefaultEnrollmentLayout.constants';
+import { NonBundledDhis2Icon } from '../../NonBundledDhis2Icon';
+import type { PlainProps, Props } from './WidgetHeader.types';
+import { OverflowMenuComponent } from '../OverflowMenu';
+
+const styles = {
+    menu: {
+        marginLeft: 'auto',
+    },
+    icon: {
+        marginRight: spacersNum.dp8,
+    },
+};
+
+
+const WidgetHeaderPlain = ({
+    linkedStage,
+    linkedEvent,
+    orgUnitId,
+    currentPage,
+    eventId,
+    relationship,
+    relationshipType,
+    stage,
+    classes }: Props) => {
+    const { icon } = linkedStage;
+    return (
+        <>
+            {icon && (
+                <div className={classes.icon}>
+                    <NonBundledDhis2Icon
+                        name={icon?.name}
+                        color={icon?.color}
+                        width={30}
+                        height={30}
+                        cornerRadius={2}
+                    />
+                </div>
+            )}
+            <span> {linkedStage.name} </span>
+            {currentPage === EnrollmentPageKeys.VIEW_EVENT && (
+                <div className={classes.menu}>
+                    <OverflowMenuComponent
+                        linkedEvent={linkedEvent}
+                        relationshipId={relationship}
+                        relationshipType={relationshipType}
+                        orgUnitId={orgUnitId}
+                        originEventId={eventId}
+                        stageWriteAccess={stage?.access?.data?.write}
+                    />
+                </div>
+            )}
+        </>
+    );
+};
+
+export const WidgetHeader: ComponentType<PlainProps> = withStyles(styles)(WidgetHeaderPlain);
+