+++ src/core_modules/capture-core/components/WidgetTwoEventWorkspace/WidgetHeader/WidgetHeader.types.js
@@ -1,0 +1,21 @@
+// @flow
+import { EnrollmentPageKeys }
+    from '../../Pages/common/EnrollmentOverviewDomain/EnrollmentPageLayout/DefaultEnrollmentLayout.constants';
+import type { ProgramStage } from '../../../metaData';
+
+export type PlainProps = {|
+    orgUnitId: string,
+    linkedEvent: { event: string },
+    linkedStage: ProgramStage,
+    currentPage: $Values<typeof EnrollmentPageKeys> | string,
+    eventId: string,
+    relationship: string,
+    relationshipType: string,
+    stage: ProgramStage,
+|};
+
+export type Props = {|
+    ...PlainProps,
+    ...CssClasses,
+|};
+