+++ src/core_modules/capture-core/components/WidgetTwoEventWorkspace/WidgetWrapper/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { WidgetWrapper } from './WidgetWrapper.container';