+++ src/core_modules/capture-core/components/WidgetsChangelog/WidgetEventChangelog/WidgetEventChangelog.js
@@ -5,6 +5,7 @@ import { Changelog, CHANGELOG_ENTITY_TYPES } from '../common/Changelog';
 
 type Props = {
     eventId: string,
+    eventData: Object,
     dataItemDefinitions: ItemDefinitions,
     isOpen: boolean,
     setIsOpen: (boolean | boolean => boolean) => void,
@@ -12,6 +13,7 @@ type Props = {
 
 export const WidgetEventChangelog = ({
     eventId,
+    eventData,
     setIsOpen,
     ...passOnProps
 }: Props) => (
@@ -19,6 +21,7 @@ export const WidgetEventChangelog = ({
         {...passOnProps}
         close={() => setIsOpen(false)}
         entityId={eventId}
+        entityData={eventData}
         entityType={CHANGELOG_ENTITY_TYPES.EVENT}
     />
 );