+++ src/core_modules/capture-core/components/WorkingLists/TeiWorkingLists/TrackedEntityBulkActions/Actions/DeleteEnrollmentsAction/EnrollmentDeleteModal/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { EnrollmentDeleteModal } from './EnrollmentDeleteModal';